widgets: Cache event handler instead of entire jquery for each widget.
In 6349e754def8ae032618f71ec1509d8091af3811 we stopped restoring widgets
from the cache, so the only purpose of the cache remaining was to access
the event handler patched onto the widget jquery.

Now instead of the entire jquery object, we cache just the event handler
since that's the only thing we need from the cache, and do not patch it
onto the widget jquery object.