zcommand: Rename night->dark and day->light in the color scheme backend.
As a follow up for 613500bed16a9914f464de4576d94423bfb404b4, this
commit standardizes the naming of the day and night themes to light
and dark, respectively in the backend. This makes the backend
consistent with the naming used in the frontend and UI.

This also solves a regression introduced in
613500bed16a9914f464de4576d94423bfb404b4, where the frontend was sending
"/light" and "/dark" commands to the backend, but the backend was
expecting "/day" and "/night" commands.