settings: Add 'web_navigate_to_sent_message' setting.
In a2ef1d7e9376de270be7537472f937b335119e53, we made changes so
that when you send a message, your view jumps to the conversation
where you sent it.

For some users it was an improvement, few reported that it
disrupts their workflows.

This prep commit adds a setting which will be used to allow users
to decide whether to automatically go to conversation where they
sent a message.