compose: Don't jump to sent message conversation if setting disabled.
In a2ef1d7e9376de270be7537472f937b335119e53, we made changes so
that when you send a message, your view jumps to the conversation
where you sent it.

For some users it was an improvement, few reported that it
disrupts their workflows.

This commit updates the logic to NOT jump to the conversation
where you sent message if the setting
'Automatically go to conversation where you sent a message' is
disabled in 'SETTINGS / PREFERENCES > Advanced'.

We restore the old behaviour for the setting disabled case.

Fixes #30600.