popovers: Allow popover hotkey hints min-width to scale with font size.
A `min-width` is applied to the popover hotkey hints to ensure that they
are nearly square when only one character is present. This `min-width`
was previously using rem units, which did not scale with the font size
of the popover font.

This commit changes the `min-width` to use `em` units, which will scale
with the font size of the popover hotkeys.