user_card_popover: Remove unused user `undefined` check.
This was introduced in 636d4253a5c18aeda921b12754a4b151399e3437, and
since the comment said it should never happen, it felt safe enough to
remove.

The `message` argument has also been replaced with `sender_id` since
that is the only property we need from the message now. This will
become useful in adding support for these popovers to compose box where
the message object will not be present but there would be a sender_id.