message_edit: Update stream active status when moving messages.
Update the active status of new stream where the messages are moved
into, if appropriate.

Tested by deleting all messages in a stream. Running the command
to update stream status to mark it inactive. Then moving messages
into the stream to check if the status is updated correctly to active.