activity: Use icon for activity links and realm name for support view.
In the installation and integrations activity views, updates the
links to a realm's activity view to use the 'fa fa-table' icon,
instead of the realm name. Now, the realm name links instead to
the realm's support page.

Also, the integrations activity page now includes a link to a
realm's stats view.