streams: Do not call "locals()" inside loop.
There is no need to call "locals()" inside the loop for
stream permission settings. It should just be called once
to get values of all the settings passed to the endpoint.

It was fine as we only had one group permission setting
for stream, but is a good fix before we add more settings.