compose: Remove duplicate call to clear_preview_area.
This first call was added in this commit:
cb5bb4e2f997d01f0b63bf26533c9ebb9985eb83
At the time of this commit, the second call was already there.
When removing this line in this commit, it seems like the
earlier reported bug isn't present. So we can remove it, to avoid
doing extra work when sending a message.