popovers: Hide left sidebar actions popover for empty topics.
This commit changes the behavior of the topic actions popover
(displayed from the left sidebar via the `...` menu) to no
longer appear for empty topics (topics with no messages).
Instead, a message is displayed: "There are no messages in
this topic.", along with the topic visibility options and
the name of the topic.

Fixes: #32098.