footer: Fix footer items mis-aligned in Russian.
Change alignment of footer rows to flex-start a bit earlier.
`1280px` is chosen since that's the standard `max-width` for
the screens on the page and that is width around which the
massive gap before the first item on the footer is diminished which
allows us to left align the elements.