provision: Source virtualenv for login shells conditionally.
This commit conditionally adds the source command for
virtualenv in the user's bash profile only when the
system is virtualized.

Checking WSL_DISTRO_NAME in the env vars
is the cheapest option in case of checking
whether the system runs as WSL VM since
it is an in memory operation.
Hence it is checked for first in `is_wsl_instace`.

Fixes #15029.