settings: Use only `Filter` as the placeholder text.
There are cases where the placeholder text overflows outside of the
input box. In the settings panels, all these filter boxes are on the
same row as a subheading on the left side. So the x in `Filter x` is
usually the subheading title, making it redundant to mention the x part.
We have not modified the aria-labels since it might still be helpful for
assistive technologies to have the whole `Filter x` part.
https://chat.zulip.org/#narrow/channel/101-design/topic/setting.20filter.20width