polls: Do not use pixel height for poll title.
Poll title was set to 18px which was leading to the title looking
smaller than the options in 20px font size. Removing that explicit
font-size declaration will make the h4 font-size fall back to 1.1em and
thus that title will scale properly to font-size changes. At 16px, the
font-size for the title was 18px initially, now it is 17.6px which
should not be that big of a change.