groups: Fix code to show newly added settings in permissions panel.
The newly added "can_create_write_only_bots_group" and
"can_create_bots_group" were not added in "Permissions" panel
of groups UI if the group has that permission and this commit
fixes it.

We also add a function called in ui_init.ts to make sure that
the settings data defined in settings_config.ts used for
"Permissions" panel contains all the group settings.