group-settings: Refactor code to fix live update of permissions panel.
This commit refactors the code so that we can correctly update the
permissions panel when a group gains permission. The changes done are-

- Extracted templates for subsection for stream and group settings
because we might need to add complete subsection to DOM if group has
no existing permission for a particular stream or group.

- Also, updated get_group_assigned_realm_permissions to return all
subsection objects even when group has no assigned permission for
a subsection because we now have headers for all the subsection
in DOM but just hide them so that order can be maintained when
a setting of a subsection is added during live update.