compose_recipient: Early return in on_hidden_callback if no item click.
Previously, the topic / direct message recipient input box was always
focused when the compose select recipient dropdown got closed, even if
the user clicked elsewhere. This resulted in unwanted trigger of
onFocus event listener etc.

This adds an early return to ensure focus updates only when an item
is clicked.