typeahead: Add showOnClick option to customize input click behavior.
Previously, via commit 0dc7f5f91fab1a8e311ef306e47ce2e0e6dad61b, the
logic for the click action on the input field was added to the typeahead
to aid with the interaction when the user changes the position of the
cursor — targeted at the compose box where there could be multiple
instances of typeahead.

This commit adds an option to disable the rendering of typeahead when
the user navigates within the input element via clicks — suitable for
single input elements where we don't expect multiple typeaheads.