alerts: Remove --base-font-size-px and allow font-size to be inherited.
This is a follow-up commit to 7739a10c8888a04567ee4aeed7aed29cb6386668,
which had set the font-size for the alert popups to --base-font-size-px.

This removes that line of code, as it was pointed out by Tim Abbott
that the web/styles/alerts.css file was also being shared with portico
through the web/src/bundles/common.ts bundle, but the
--base-font-size-px variable doesn't exist there.

Removing the font-size property from the alert box doesn't affect its
styling since it still inherits the `--base-font-size-px` from the body
in the Zulip Web App.