recent_view: Avoid rendering rows before user scrolls to them.
This is just a bug fix based on observation that we would only
be updating rows that are rendered and not all of them. This can
potentially lead to rows being rendered twice but is hard
to observe since newly updated topics usually are the top of the
recent view.