migrations: Update rename_general_chat_to_empty_string_topic migration.
Earlier, the migration code was replacing the occurrence of "general
 chat" as topic name in the database with `""`. This resulted in an
error which permanently broke `/near/` URL links to existing topics
named "general chat".

This commit updates the migration approach to instead move messages
from "general chat" to `""`. An edit history entry is added.

It results in the same expected behaviour and fixing that bug.