streams: Remove require_active in access_stream code path.
This was confusingly doing an assertion about the subscription being
active, not the channel. We could rename it to
require_active_subscription. But it was only passed with a non-default
value in 64924b46cc0166f185f2c2d2e333f0a058a44557, and that call was
removed in d12e083eff8aea1be88b50ddc320e58ca4678007.