message-edit-tooltip: Use message_container.modified to show tooltip.
When we build the message_container for the message list view, we
get a boolean value for showing the edit notice (saving, moved, or
edited) on the message, so we now use that for whether we show the
associated message edit tooltip as well.

Also, if the message is modified, then we either have a last edited
timestamp from the server or from a local edit, so it shouldn't be
undefined.