message-edit-history: Show original channel/topic for move only history.
If the organization only allows viewing message move history, in
the message edit history overlay for the initial message, we display
the original topic and channel of the message instead of the original
message content.

This way the overlay content is consistent with the organization
setting for viewing message edit history, and if the message content
has been edited, that original content is not viewable to other
users.