upgrade: Remove explicit python3-yaml install step.
The python3-yaml dependency was added at install time in 21858f6e22b9
in Zulip Server 4.0, and this workaround was added in f4ca340e5ee9,
also in 4.0.  Since we can only directly upgrade from 5.0 or later,
the dependency is guaranteed to be installed already, by one or the
other of those ways. Remove this workaround.