Revert "css: Remove unused transition property."
This reverts commit 65367246c770edd1d26d661b8b367edc5a60ee60.

The commit message is factually incorrect, and the change broke our
fade-out transitions.  visibility can be transitioned, and there is a
specific rule for what that means.  This is critical for transitioning
opacity in the fade-out direction: if you fail to also transition
visibility, then the element disappears immediately regardless of
opacity.

https://developer.mozilla.org/en-US/docs/Web/CSS/visibility#interpolation

Signed-off-by: Anders Kaseorg <anders@zulip.com>