nginx: Serve full app from localhost.
Some deployments choose to wrap Zulip's nginx in an outer proxy -- for
example, to do custom TLS termination.  In such deployments, the outer
proxy is routing to `127.0.0.1:80`; 9548cab9167a breaks these
configurations, as it switches the `127.0.0.1:80` listener to only
serving `/api/internal/` paths.

Switch to serving the whole application over `127.0.0.1:80`.