message_row: Remove tag-level filtering in the date row timerender CSS.
This commit improves the CSS matching performance by avoiding tag-level
filtering in the date row CSS. Instead of using the `span` tag we use
the ".timerender-content" class to filter the same time/date elements.

Since this class is added by `timerender.render_date()` to the span
element inside date_row on every instance, we can use it with assurance.