demo-orgs: Separate event listeners from convert org modal.
Updates handle_demo_organization_conversion to only set up the
event listeners for the current convert organization banner in
the organization settings overlay for admins.

Moves code for launching the modal to convert the demo organization
to a permanent organization to do_convert_demo_organization.

Prep for adding a button to the navbar alert banner for demo
organization owners that will also launch the modal to convert
the organization.