integration-docs: Clarify the download-python-bindings macro.
Replaced "server" with "system" so that the macro can be used with all
integrations with integration scripts, irrespective of whether it needs
to be run from a server or a user's machine.

Replaced "bot" with "integration script" to ensure that the integration
script is not confused with the Zulip bot user that is created using the
UI in the instruction above this instruction.