groups-ui: Do not open group in right panel when leaving.
Previously, when leaving the group by clicking "checkmark"
icon in group row, group details were opened in right panel
as well due to event propagation.

We keep the behavior of opening the group in right panel
after joining the group by clicking on "+" icon in the
group row.

For narrow widths, left panel is kept open while joining
or leaving the group, same like we do for stream settings.