groups-ui: Fix exiting from creation UI for mobile widths.
When clicking on "Cancel" button in the footer in stream
creation UI, left panel should be shown for mobile widths.
Previously, it resulted in a buggy view showing both left
and right panels.

This change makes the behavior consistent with what we
have for streams UI.

There is no change in behavior for wider screens.