integrations: Remove Nagios fixture used for example screenshot.
Nagios is not a webhook integration, and the fixture is only used to
generate the example screenshot for the doc.

Support for generating the Nagios example screenshot will be added in a
later commit, in a different format and system. Removing this fixture
allows removing the current system logic, without raising errors.