help: Remove "Start a video call" mobile tab.
The ability to add a link to a video call is not yet implemented
for the initial launch of the Flutter mobile app, so we remove
the legacy React Native app instructions.

Deletes static/images/help/mobile-video-icon.svg as it is no
longer used.

This feature is tracked in:
https://github.com/zulip/zulip-flutter/issues/1000

Part of #34748.