help: Remove "Status and availability" mobile tabs.
The ability to view, set, edit a user's status, or to toggle invisible
mode are not yet implemented for the initial launch of the Flutter
mobile app, so we remove the legacy React Native app instructions.

Deletes static/images/help/mobile-x-icon.svg as it is no
longer used.

This feature is tracked in:
https://github.com/zulip/zulip-flutter/issues/197
https://github.com/zulip/zulip-flutter/issues/198
https://github.com/zulip/zulip-flutter/issues/1578

Part of #34748.