support: Update view for placeholder realms.
When a realm is a placeholder realm due, i.e., the realm redirects
to another URL due to previously changing the realm's subdomain,
we only show the redirect information.

There is a copy link that for the redirect URL that can be pasted
into the search bar to get to the support view for the actual
realm.