api_docs: Update the changelog for feature 318.
This changelog for feature 318 was added in this commit:
23ffc3474dd64bc4792f6f12727346d706612ae9. The changelog stated that it
added a new field called `config_options` to the
`realm__incoming_webhook_bots` field, but in reality, the it renamed an
existing field called `config`.

This fixes the changelog for feature 318 and its changes note in the API
documentation to correctly describe the change.