users: Fix `check_valid_bot_config` bug.
Currently `check_valid_bot_config` checks if `config_options` has a
falsy value to determine whether an integration is invalid or not. But,
what it actually does is check whether an exists in WEBHOOK_INTEGRATIONS
and it has at least one WebhookConfigOption.

It should instead, check if `config_options` is `None` to determine
whether an integration is invalid or not.