topic_popover: Show loading indicicator on topic resolve/unresolve.
This commit shows a loading indicator in the recipient bar when the
user resolves or unresolves a topic from the topic actions menu.

Since we do not have a unresolve icon button in the recipient bar
following commit d24883c56a68a7bc13fc7d3bcb84119ddb0d1584, we
first render a icon button with a placeholder icon in the recipient bar,
and then show the loading indicator when the user clicks on the
"Mark as unresolved" option in the topic actions menu.