drafts-ui: Fix banner shown on deleting drafts.
Banner shown on deleting drafts was broken because "banner-wrapper"
class was missing in container element that led to some CSS not
being applied.

Also, needed to change the "text-align" CSS property to override
the "center" value set for overlay header text because "banner-label"
flex element is allowed to grow and take more width after recent
changes in 777da20.